--- builtins/amp-layout/build.js
@@ -1,38 +1,0 @@
-/**
- * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {parseLayout, Layout} from '../../src/layout';
-import * as dom from '../../src/dom';
-
-/**
- *
- * @param {!Document} document
- * @param {!Element} element
- */
-export function buildDOM(document, element) {
-  const layout = parseLayout(element.getAttribute('layout'));
-  if (layout == Layout.CONTAINER) {
-    return;
-  }
-
-  const container = document.createElement('div');
-  dom.applyFillContent(container);
-
-  dom.getRealChildNodes(element).forEach((child) => {
-    container.appendChild(child);
-  });
-  element.appendChild(container);
-}
